move client to new module, add shop and shop listings methods
